api_server/requirements.txt
acres==0.5.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.5
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
fastapi==0.135.1
fastembed==0.7.4
filelock==3.25.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.12.19
frontend==0.0.3
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.66.0
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.17
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-ollama==1.0.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.12
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6.1
nibabel==5.4.0
nipype==1.11.0
numpy==2.4.2
ollama==0.6.1
onnxruntime==1.24.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
pathlib==1.0.1
pillow==11.3.0
portalocker==3.2.0
protobuf==7.34.0
prov==2.1.1
puremagic==1.30
py_rust_stemmers==0.1.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyxnat==1.6.4
PyYAML==6.0.3
qdrant-client==1.17.0
rdflib==7.6.0
redis==7.2.1
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
traits==7.1.0
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
debugpy==1.8.1